message d'erreur suppression

+++ Controller/AnnonceController.php
@@ -3,6 +3,7 @@
 namespace Controller;
 
 use model\Annonce;
+use model\AppConstant;
 use service\AnnonceService;
 
 require '../service/AnnonceService.php';
@@ -19,7 +20,31 @@ class AnnonceController
     // La fonction permettant de supprimer une annonce
     public static function deleteAnnonce(): void
     {
-        AnnonceService::deleteAnnonce();
+        // Récupération de l'utilisateur connecté
+        $userConnect = getElementInSession(AppConstant::USE_ID_SESSION_KEY);
+
+
+        if ($userConnect) {
+
+            // Récupération de l'id du user connecté
+            $userConnectId = $userConnect->getUseId();
+
+            // récupération de l'id du créateur de l'annonce
+            $userAnnonceId = getElementInRequestByAttribute("userAnnonceId");
+
+            // Comparaison des deux ids
+            if ($userConnectId == $userAnnonceId) {
+                AnnonceService::deleteAnnonce();
+            } else {
+                $_SESSION["errorDeleteAnnonce"] = "vous n'avez pas la possibilité de supprimer cette annonce";
+                header(AppConstant::$HEADER_LOCATION_LABEL . AppConstant::$DETAILS_ANNONCE_LOCATION_LABEL);
+            }
+
+        } else {
+
+            header(AppConstant::$HEADER_LOCATION_LABEL . AppConstant::$DETAILS_ANNONCE_LOCATION_LABEL);
+            $_SESSION["errorDeleteAnnonce"] = "aucun utilisateur connecté";
+        }
     }
 
     // La fonction permettant de supprimer une annonce
